Updating readability. Added 3rd item.
Replaced instances of 'inventory_item' with 'item'. Calling everything an inventory_item is redundant since this is an inventory management system, therefore, all items are inventory_items.

This would NOT be the case if we were keeping track of stuff in a store that is NOT for sale. Such as window cleaner, number of mops, etc.

Added Skateboard to reflect price can appear as i32 when actually f32 from our InventoryItem struct.