Updated function to return string.
By editing total_value_of_inventory to return a string, our test is able to print the result of the math problem.

total_value_of_inventory only needs to print a string for our needs. At this point, we do not need this function to write the result to a file, just to show us the value of our total inventory.